Move more parts out of home.nix
This is mainly so I can reuse more of it later. This also moves
herbstluftwm config into home/hlwm. polybar is next, as soon as I get
around to cleaning up that mess of a config. For now, it continues
to be its own thing that needs to be carefully managed, unfortunately.

This also removes alacritty from the default setup. Its config is
still around in case I need to switch for some reason, but I find
kitty to be friendlier overall still.