Add helix, typst, add ncrandr param to hlwm
Add helix and typst flakes for personal use. Helix config still needs
to be added and then replaced with home-manager config for it, but
this is a start. Still playing with typst so installing it.

Add an ncrandr param to the hlwm module instead of referencing it
directly. This avoids the need to figure out where to put it in pkgs.

Move afmt and ncrandr to nixpkgs-unstable.