Add Helix config, remove mise
Add Helix config to home/helix. This is a direct mostly mechanical
translation of my normal Helix configs, just dumping them into nix
and changing them from toml.

Remove mise because it recently added an external dependency for
usage info, doesn't let you turn it off, and causes fish to print
completion errors. I assume this will be fixed at some point, but
it's bad UI design and frustrates me, so I won't use it. Probably
not as much reason to have it in Nix in the first place.