Add polybar config
Add polybar config to the hlwm module. Originally planned on using
the polybar config in home-manager for this, but it tries to run it
through systemd, which is more work than I want to figure out right
now when using herbstluftwm reloads to restart it. So, instead, this
bundles it into hlwm and runs it manually by tying its lifespan to
the next herbstluftwm reload.

The config INI generation is messy, but works. I can always fix it
up later.