Add git-mr script to git-tools
This adds git-mr to git-tools, which is what I use most of the time for
opening pull / merge requests. Some of this is negated these days by
the `gh` tool GitHub provides, which I've also added to package sets.

This also cleans up the git-tools packaging a bit so it's less
patch-y. Now the only thing that really needs to be known is the source
and any runtime dependencies of the source.