Organize modules under home and nixos
Organize (most) modules under the home/ and nixos/ directories. The
main exception to this is currently the fish/ stuff, which is the
oldest and I want to spend some more time cleaning up before I move it
into home/.

The majority of this work is just splitting all the file imports off
into either members of nixosModules or members of homeManagerModules
for the root flake. This is just to avoid having to know where a file
needs to be referenced from relative to the importing file, since
that's a headache.