Fix ncrandr handling modes immediately after the EDID
If mode lines were found after the EDID, they would get included in the
EDID and hashed, resulting in a different hash than normally found. The
fix for this is just a hack: only take the property lines following a
property name that begin with two tabs. This isn't robust, but it works
for now.