Add a bunch of packages of my own
These are programs that I don't really want to add flakes to (many are
years old and I really only want them for scripts). Might still remove
the sub-flakes for other things and just package stuff normally as
well since the sub-flakes are a pain in the neck at times (see README
for quirks).