Add git and helix to dev shell
Add git and helix to the dev shell. This is standard-packaged helix,
not my fork, so it should work out of the box while getting a machine
set up. (Otherwise it would take way too long to get a dev shell
started -- the cost of running the crystal compiler is enough right
now).