Use NixOS 24.05
This re-pins most config back to 24.05. No changes to state
version. Want to get Sirin updated since it's the furthest behind
(it was working, so didn't feel any pressing need to do a lot to it),
but don't want to pin it to unstable just yet.

Also update the kernel to 6.9 since 6.7 is EOL. Shame I can't use that
for work.

Move services.xserver.libinput to services.libinput since this changed
in 24.05. No major problems there.

Swap ifEnable out for a regular if. It looks like the second parameter
of ifEnable has to be evaluated to determine the type of the false
result, which isn't desirable since not all configs include nixGL
overlays.