Update Helix to 24.07
Just updates Helix to 24.07. Tried to move my patch directly into
this flake and patch it, but Helix being wrapped in a way that doesn't
appear to support overrides makes it a little difficult right now. So,
I don't get to switch back to the main repo for the flake yet, but at
least 24.07 doesn't require extra patches now.

Also noticing that piping stuff in 24.07 to fmt is a little different
around handling the last newline. May need to update afmt to account
for this.