Switch Helix flake back to main repo
Took me a bit of reading the flake to figure out there's a way to
regenerate the wrapper from a patched build, via the `wrapper`
passthru, so using that now and adding my patch and the one I
cherry-picked to fix pipes. This should hopefully make it unnecessary
to continue maintaining patched branches of Helix itself.