Fix kakoune-lsp config, add kak-session-name
Fix the kakoune-lsp config, mostly in that it's missing if you use it
as a regular plugin. This is probably assumed to be set up manually,
but I'm lazy, so I just yoinked it out of the kakoune-lsp source.

Add my kak-session-name plugin because I prefer session names over IDs
(and a little easier with fish completing session names on tab).